requirements.txt
openai
pillow
geopy